Add code in Task.java to handle blank description
Sub-classes of Task class accepts task description that is empty or just
only whitespaces. Realistically, this means the task description is
empty and the tasks should not be created in the first place.

Let's Add a condition to ensure that task created has a non-empty task
description, i.e. task description should contain non-whitespace
characters.

This is to improve user experience and functionality.