Update Command and UI to follow SRP
The UI and Command prints out the results of the command.
This is not ideal as the GUI would not work properly.

Let's change the methods in these classes to return strings of the
outputs.

This will allow the output to be easily passed to the GUI while
maintaining usability of the Yap chatbot on command line.