Change the date and time syntax
For event classes, the default date and time syntax to follow is
quite awkward and complicated.

This can result in a bad user experience.

Let's change the date and time syntax to something that is more
user-friendly.

This way, dates and times will be presented in a more user-readable
format.