Add support for error-formatted responses
Responses are passed to the UI as strings. This
makes it difficult for errors to be passed
and specially formatted.

Create new ChatResponse class to encapsulate
the response string as well as response-related
information - in this case whether it is an
error message. So when it is passed to the UI
the UI can determine what kind of formatting
needs to be applied to the response message.