Add FixedDurationTask class
Currently, there exists no class for an unscheduled task with a fixed
duration, which might be useful for users who want to create tasks
without a specific date in mind.

Let's create a class for this, named FixedDurationTask, which extends
from a Task to inherit its useful properties and methods. The duration
will be in the form of hours.

This is so that the data for this type of task can easily be saved and
read from.