Add handlers for FixedDurationTask
After adding FixedDurationTask class, we need to modify the codebase to
start integrating the FixedDurationTask so that users can manage this
type of task.

We do this by adding the appropriate commands, and methods, in a manner
similar to integrating other tasks.

We do this to make use of the extensibility of the codebase, and
maintain consistencyAdd handlers for FixedDurationTask

After adding FixedDurationTask class, we need to modify the codebase to
start integrating the FixedDurationTask so that users can manage this
type of task.

We do this by adding the appropriate commands, and methods, in a manner
similar to integrating other tasks.

We do this to make use of the extensibility of the codebase, and
maintain consistency.