extract string formatting as a FormattedString utility class
There is a need for classes to get the unformatted version of a string.
A FormattedString class would also enable greater flexibility when
formatting strings in the future, for example applying many formats to
a string.

Let's extract a FormattedString class that can apply red and
yellow-background to strings for now.
The FormattedString class acts like the StringBuilder class in Java
which allows chaining operations on it.