Add feature to find nearest x free days
The Free command allows the user to find the nearest day where the next
x days are not allocated to Tasks. e.g. If there is an Event from
23-09-2024 to 25-10-2024, and the user wants to find the nearest 4 free
days from today (26-09-2024), then the result will be 26-10-2024.

This command currently only works with Days because there is only
implementation of LocalDate and not LocalDateTime. Let's work towards
making the Free command compatible with other options such as Hours.