Add and edit JavaDoc in Command classes
Since the EditCommand class was newly added, there are no JavaDoc
comments in any of its methods. The execute() method in Command class
was modified to return a String, so the JavaDoc comments for the other
classes are also not updated.

Not updating the JavaDoc comments properly would make it hard for other
developers seeing the code to understand its functionalities and how to
use the methods.

Let's put in updated JavaDoc comments for the Command class and its
subclasses.