Add JavaDoc comments and edit editTask() method in TaskList class
There are no JavaDoc comments in the editTask() method. And in the
editTask(), the wrong task is being edited since the taskList index
starts with 0 instead of 1.

Not updating the JavaDoc comments properly would make it hard for other
developers to understand how to use editTask().

Let's put in updated JavaDoc comments for the TaskList class and correct
the index for the task to edit in editTask().