Add JavaDoc comments and edit changeDueDate() in Deadline class
There are no JavaDoc comments in the changeDueDate() method. And in the
method, the DateTime format is incorrect since the month and hour
should be capitalized.

Not updating the JavaDoc comments properly would make it hard for other
developers to understand how to use changeDueDate(). Formatting the
date and time incorrectly will throw an exception since it can't
be parsed.

Let's put in updated JavaDoc comments for the Deadline class and correct
the DateTime format in changeDueDate().