Add JavaDoc comments and edit constructor in Event class
There are no JavaDoc comments in the changeStartTime() and
changeEndTime() methods. And in the constructor, the DateTime format is
incorrect since the month and hour should be capitalized.

Not updating the JavaDoc comments properly would make it hard for other
developers to understand how to use changeStartTime() and
changeEndTime(). Formatting the date and time incorrectly will throw an
exception since it can't be parsed.

Let's put in updated JavaDoc comments for the Event class and correct
the DateTime format in the constructor.