Fix bug related to todo event parser
Bug occurred where input "todoexample" would result in an event of
"xample" being added as there was code to remove the first 5
characters of the input string.
Now the bug is fixed by changing up the code to detect if there
is no whitespace between the command todo and the description and
will throw an appropriate exception.